requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
jinja2==3.1.5

# LLM APIs
openai==1.59.6
anthropic==0.42.0
google-genai==1.5.0

# Configuration
pydantic-settings==2.7.1
python-dotenv==1.0.1

# Database
aiosqlite==0.20.0

# Audio / Speech (free, no API key)
pydub==0.25.1
SpeechRecognition==3.12.0
gTTS==2.5.4

# Utilities
aiofiles==24.1.0
